refactor: Move enum to constants
