+++ apps/web/src/local-studio/constants/index.ts
@@ -1,0 +1,1 @@
+export * from './EnvironmentEnum';
